refactor: comment 쿼리 위치이동
