+++ src/components/post/postDetail/CommentOptions.tsx
@@ -1,4 +1,4 @@
-import { useDeleteComment } from '@/hooks/queries/byUse/useCommentMutation';
+import { useDeleteComment } from '@/hooks/queries/comments/useCommentMutation';
 import { Comment } from '@/types/postDetailTypes';
 import React, { useEffect, useMemo, useRef, useState } from 'react';
